pyproject.toml
[project]
name = "mcp-terminal"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ollama>=0.5.1",
    "openai>=1.99.6",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "sqlalchemy>=2.0.42",
    "mypy>=1.17.1",
    "altair[all]>=5.5.0",
    "vega-datasets>=0.9.0",
    "phidata>=2.7.10",
    "agno>=2.1.0",
    "huggingface-hub>=0.35.3",
    "fastmcp>=2.12.4",
    "fastapi[standard]>=0.119.0",
    "googlesearch-python>=1.3.0",
    "pycountry>=24.6.1",
    "google-genai>=1.43.0",
    "ddgs>=9.6.1",
    "fuzzywuzzy>=0.18.0",
    "python-levenshtein>=0.27.1",
    "fastapi-keycloak>=1.1.0",
    "pydantic>=2.12.0",
]

[dependency-groups]
dev = [
    "ruff>=0.12.8",
]

[tool.ruff]
line-length = 120
